requirements.txt
pydub
moviepy
streamlit
python-dotenv
google-generativeai
speechrecognition